clean up routes
